@@ -3,6 +3,11 @@ import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { IPelis } from '../models/IPelis.interface';
+import { conversation} from '../_interfaces/user';
+import firebase from 'firebase/app';
+import "firebase/auth";
+import 'firebase/database';
+import { Storage } from '@ionic/storage-angular';
 
 @Injectable({
   providedIn: 'root'
@@ -10,7 +15,10 @@ import { IPelis } from '../models/IPelis.interface';
 export class PeliService {
   private url: string = '';
 
-  constructor(private http: HttpClient) { }
+  db: firebase.database.Database
+  constructor(private http: HttpClient, public storage: Storage) {
+    this.db = firebase.database();
+   }
 
   searchMovies(title: string, type: string){
     this.url = `http://www.omdbapi.com/?s=${encodeURI(title)}&type=${type}&apikey=d84c07af`;
@@ -21,4 +29,7 @@ export class PeliService {
   getDetails(id: string){
     return this.http.get<IPelis>(`http://www.omdbapi.com/?i=${id}&plot=full&apikey=d84c07af`);
   }
+  async create(conversation: conversation) {
+    await this.db.ref(`favorite/${conversation.uid}/${conversation.timestamp}`).set(conversation);
+  } 
 }
